requirements.txt
accelerate==1.5.2
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Django==4.2.20
django-cors-headers==4.7.0
djangorestframework==3.16.0
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
ftfy==6.3.1
huggingface-hub==0.29.3
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
open_clip_torch==2.31.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==3.20.3
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.13.1
seaborn==0.13.2
sentencepiece==0.2.0
six==1.17.0
sqlparse==0.5.3
sympy==1.13.1
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.13.0
tzdata==2025.2
ultralytics==8.3.94
ultralytics-thop==2.0.14
urllib3==2.3.0
wcwidth==0.2.13
zipp==3.21.0
